[AN] refactor: 검색 관련 공통 컴포넌트 분리 및 코드 구조 개선 #795
*   [AN] refactor: CustomLinearProgressBar 패키지 위치 이동 (search/category/component -> search/component)
*   [AN] refactor: SearchedHearitItem 및 SearchResultItem에서 사용되는 공통 UI 로직을 HearitTitle, HearitMetaRow, HearitProgressBar 컴포저블로 분리
*   [AN] refactor: SearchDetailTopBar 내 중복되는 검색 실행 로직을 performSearch 함수로 추출 및 개선
